requirements.txt
asgiref==3.4.1
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.10.0
distro==1.6.0
Django==3.0
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.10.0
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
num2words==0.5.10
numpy==1.21.2
opencv-python==4.5.3.56
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
pdf2image==1.16.0
pdfkit==0.6.1
pdfminer.six==20211012
Pillow==8.3.2
pycparser==2.20
PyJWT==2.2.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
tabula-py==2.3.0
tabulate==0.8.9
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
